requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
python-docx
langchain
langchain_openai 
langchain_core
python-dotenv
langserve
fastapi
uvicorn 
sse_starlette